+++ packages/d2-crud-plus-example/src/business/modules/example/views/form/select/crud.js
@@ -27,7 +27,7 @@ export const crudOptions = (vm) => {
       {
         title: 'ID',
         key: 'id',
-        width: 90,
+        width: 50,
         form: {
           disabled: true
         }
@@ -84,6 +84,11 @@ export const crudOptions = (vm) => {
             click (event) {
               console.log('点击事件', event)
             }
+          },
+          props: {
+            elProps: {
+              disableTransitions: true
+            }
           }
         }
       },
@@ -111,6 +116,11 @@ export const crudOptions = (vm) => {
             value: '2' // 默认值
           },
           helper: 'dict.getData可以覆盖全局配置的getRemoteDictFunc'
+        },
+        component: {
+          props: {
+            type: 'text' // 不使用tab，纯文本展示
+          }
         }
       },
       {