+++ packages/d2-crud-plus-example/src/business/modules/example/views/form/slot/index.vue
@@ -14,7 +14,9 @@
         ref="d2Crud"
         v-bind="_crudProps"
         v-on="_crudListeners"
-        @custom-emit="customEmit">
+        @custom-emit="customEmit"
+        @custom-dialog="customDialog"
+        @row-custom="customDialogSave">
 
       <div slot="header">
 
@@ -24,7 +26,7 @@
           <template slot="slotExampleSearchSlot" slot-scope="scope">
             <el-input ref="slotExample" v-model="scope.form['slotExample']" placeholder="blur有事件触发" @blur="inputBlur('search')"/>
           </template>
-          <el-button slot="suffix"  size="small" type="success" @click="addRow"><i class="el-icon-plus"/> 新增</el-button>
+          <el-button slot="suffix"  size="small" type="success" @click="customAddDialog"><i class="el-icon-plus"/> 自定义表单对话框</el-button>
 
           <crud-toolbar :compact.sync="crud.pageOptions.compact"
                         :columns="crud.columns"
@@ -138,6 +140,67 @@ export default {
     },
     handleDialogOpened ({ mode, form, template, groupTemplate }) {
       console.log('dialog opened:mode=', mode, 'form=', form, 'template=', template, 'group=', groupTemplate)
+    },
+    /**
+     * 自定义添加对话框
+     * @row-custom=监听对话框确认按钮事件
+     **/
+    customAddDialog (context) {
+      console.log('context,', context)
+      // 打开自定义模版的对话框
+      this.getD2Crud().showDialog({
+        mode: 'custom', // 当前打开模式,可选项[add,edit,view]
+        template: {
+          key2: {
+            title: '添加字段1',
+            key: 'key2'
+          }
+        },
+        rowIndex: null,
+        modeContext: {
+          test: '这里是modeContext----add'
+        }
+      })
+    },
+    /**
+     * 自定义编辑对话框
+     * @row-custom=监听对话框确认按钮事件
+     **/
+    customDialog (context) {
+      console.log('context,', context)
+      // 打开自定义模版的对话框
+      this.getD2Crud().showDialog({
+        mode: 'custom', // 当前打开模式,可选项[add,edit,view, 还可以自定义任意字符串]
+        rowIndex: context.index, // 编辑或查看时，哪一行的数据
+        template: {
+          key1: {
+            title: '字段1',
+            key: 'key1'
+          }
+        },
+        modeContext: {
+          test: '这里是modeContext---edit'
+        }
+      })
+    },
+    /**
+     * 自定义对话框确定事件
+     * @param data
+     * @param done
+     */
+    customDialogSave (data, done) {
+      console.log('自定义对话框确认', data)
+      this.$alert(JSON.stringify(data), '保存的数据')
+      console.log('modeContext', data)
+      // 按钮loading
+      this.crud.formOptions.saveLoading = true
+      setTimeout(() => {
+        done() // 关闭对话框
+        this.crud.formOptions.saveLoading = false
+      }, 1000)
+    },
+    doDialogClosed (context) {
+      console.log('dialog closed', context)
     }
 
   }