+++ packages/d2-crud-plus-extends/src/tree-selector/lib/tree-selector.vue
@@ -2,10 +2,11 @@
   <div class="d2p-tree-selector">
     <div  class="el-cascader el-cascader--default" :class="{'is-disabled':disabled}" @click="openDialog">
       <div class="el-input el-input--default el-input--suffix" :class="{'is-disabled':disabled}" >
-        <el-input ref="reference" :disabled="disabled" />
+        <el-input ref="reference" :disabled="disabled"  />
         <span class="el-input__suffix">
-          <span class="el-input__suffix-inner"><i class="el-input__icon el-icon-arrow-down" @click="openDialog"/>
-        </span>
+          <span class="el-input__suffix-inner">
+            <i class="el-input__icon el-icon-arrow-down" @click="openDialog"/>
+          </span>
         </span>
       </div>
       <div class="el-cascader__tags" ref="tags">
@@ -13,10 +14,11 @@
           <el-tag
             v-for="item in selected"
             :key="getValueKey(item)"
-            :closable="false"
+            :closable="clearable"
             :size="collapseTagSize"
             :hit="false"
             type="info"
+            @close="itemClosed(item)"
             disable-transitions>
             <span class="el-select__tags-text">{{ getValueLabel(item) }}</span>
           </el-tag>
@@ -24,21 +26,21 @@
       </div>
     </div>
     <el-dialog custom-class="d2p-tree-selector-dialog"
-      title="选择"
+      :title="dialogTitle"
       :visible.sync="dialogVisible"
       width="30%" append-to-body>
       <div class="tree-wrapper">
         <div v-if="treeFilter" class="filter-bar" style="padding-bottom: 20px">
           <el-input
             prefix-icon="el-icon-search"
-            placeholder="输入关键字进行过滤"
+            :placeholder="filterPlaceholder"
             v-model="filterText" size="small" >
           </el-input>
         </div>
 
         <div class="tree-body">
           <el-tree
-            :data="data"
+            :data="_options"
             @check-change="handleCheckChange"
             @current-change="handleCurrentChange"
             :filterNodeMethod="filterNode"
@@ -50,8 +52,8 @@
       </div>
 
       <span slot="footer" class="dialog-footer">
-          <el-button @click="dialogVisible = false">取 消</el-button>
-          <el-button type="primary" @click="selectSubmit">确 定</el-button>
+          <el-button @click="dialogVisible = false">{{cancelText}}</el-button>
+          <el-button type="primary" @click="selectSubmit">{{confirmText}}</el-button>
         </span>
     </el-dialog>
   </div>
@@ -64,7 +66,7 @@ import log from '../../utils/util.log'
 // 树形选择组件，需要import {D2pTreeSelector} from 'd2p-extends'
 export default {
   name: 'd2p-tree-selector',
-  mixins: [d2CrudPlus.inputBase],
+  mixins: [d2CrudPlus.input, d2CrudPlus.inputDict],
   props: {
     // 值
     value: {
@@ -74,6 +76,23 @@ export default {
     filter: {
       type: Function,
       require: false
+    },
+    // 过滤的placeholder
+    filterPlaceholder: {
+      type: String,
+      default: '输入关键字进行过滤'
+    },
+    dialogTitle: {
+      type: String,
+      default: '选择'
+    },
+    cancelText: {
+      type: String,
+      default: '取消'
+    },
+    confirmText: {
+      type: String,
+      default: '确定'
     },
     // 树形组件节点过滤，可以配置elProps.filterNodeMethod ，覆盖默认的过滤方法
     treeFilter: {
@@ -96,6 +115,13 @@ export default {
     elProps: {
       type: Object
     },
+    /**
+     * 是否可以清除
+     */
+    clearable: {
+      type: Boolean,
+      default: true
+    },
     // 数据字典配置
     dict: {
       type: Object,
@@ -106,7 +132,6 @@ export default {
     return {
       currentValue: undefined,
       collapseTags: false,
-      data: [],
       selected: [],
       dialogVisible: false,
       filterText: undefined
@@ -116,7 +141,7 @@ export default {
     // if (this.dict) {
     //   this.dict = d2CrudPlus.util.dict.mergeDefault(this.dict, true)
     // }
-    this.initData()
+    // this.initData()
   },
   computed: {
     _elProps () {
@@ -141,23 +166,23 @@ export default {
     }
   },
   watch: {
-    value (value) {
-      this.$emit('change', value)
-      this.dispatch('ElFormItem', 'el.form.blur')
-      this.setValue(value)
-    },
     filterText (val) {
       this.$refs.elTree.filter(val)
     }
   },
   methods: {
-    initData () {
-      d2CrudPlus.util.dict.get(this.dict).then(ret => {
-        this.$set(this, 'data', ret)
-        this.setValue(this.value)
-      })
+    // initData () {
+    //   d2CrudPlus.util.dict.get(this.dict).then(ret => {
+    //     this.$set(this, 'data', ret)
+    //     this.setValue(this.value)
+    //   })
+    // },
+    onDictLoaded () {
+      log.debug('onDictLoaded', this.dict, this.value)
+      this.setValue(this.value)
     },
     setValue (value) {
+      log.debug('setValue:', this.currentValue, this.value, this._options)
       if (this.currentValue === this.value) {
         return
       }
@@ -169,14 +194,18 @@ export default {
       if (!(arrValue instanceof Array)) {
         arrValue = [arrValue]
       }
-      if (this.dict.getNodes) {
+      if (this.dict && this.dict.getNodes) {
+        log.debug('getNodes:', arrValue)
         this.dict.getNodes(arrValue).then(nodes => {
           this.selectedNodes(nodes, value)
         })
       } else {
         const nodes = []
+        if (this._options == null || this._options.length === 0) {
+          return
+        }
         for (const item of arrValue) {
-          const data = this.data
+          const data = this._options
           const node = d2CrudPlus.util.dict.getByValue(item, data, this.dict)
           if (node != null) {
             nodes.push(node)
@@ -254,8 +283,11 @@ export default {
     },
     selectSubmit () {
       const nodes = this.refreshSelected()
-      let values = this.formatValue(nodes)
       this.dialogVisible = false
+      this.doValueInputChanged(nodes)
+    },
+    doValueInputChanged (nodes) {
+      let values = this.formatValue(nodes)
       this.resetInputHeight()
       if (!this.multiple) {
         values = values && values.length > 0 ? values[0] : undefined
@@ -263,6 +295,12 @@ export default {
       this.currentValue = values
       this.dispatch('ElFormItem', 'el.form.blur')
       this.$emit('input', values)
+    },
+    itemClosed (item) {
+      const newNodes = lodash.without(this.selected, item)
+      console.log('new value', item, newNodes)
+      this.$set(this, 'selected', newNodes)
+      this.doValueInputChanged(newNodes)
     },
     refreshSelected () {
       let nodes = null
@@ -335,6 +373,11 @@ export default {
     filterNode (value, data) {
       if (!value) return true
       return this.getValueLabel(data).indexOf(value) !== -1
+    },
+    onChange (value) {
+      this.$emit('change', value)
+
+      this.dispatch('ElFormItem', 'el.form.blur')
     }
   }
 }
@@ -344,6 +387,9 @@ export default {
   width: 100%;
   .el-cascader{
     width: 100%;
+  }
+  .is-disabled .el-tag__close.el-icon-close {
+    display: none;
   }
 }
 .d2p-tree-selector-dialog{